fixed the drawer components, so now they send props
