+++ TrafikkApp/App.js
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 
-import MyDrawer from './components/Drawer.js';
+import MainDrawer from './components/Drawer';
 
 export default function App() {
-     return (
-          <NavigationContainer>
-               <MyDrawer />
-          </NavigationContainer>
-     );
+    return (
+        <NavigationContainer>
+            <MainDrawer />
+        </NavigationContainer>
+    );
 }
