+++ TrafikkApp/package.json
@@ -17,6 +17,7 @@
         "react": "16.13.1",
         "react-native": "0.63.4",
         "react-native-gesture-handler": "^1.9.0",
+        "react-native-orientation": "^3.1.3",
         "react-native-reanimated": "^1.13.2",
         "react-native-safe-area-context": "^3.1.9",
         "react-native-screens": "^2.15.1",
