+++ TrafikkApp/styles/mainStyles.js
@@ -7,8 +7,6 @@ const styles = StyleSheet.create({
         height: '100%',
         alignItems: 'center',
         justifyContent: 'flex-start',
-
-        // backgroundColor: 'black'
     },
 });
